fixes - devide into 2 apps
