+++ BackEnd/BackEnd/settings.py
@@ -26,6 +26,8 @@ INSTALLED_APPS = [
     'rest_framework',
     'user.apps.UserConfig',
     'corsheaders',
+    'contact_info',
+    'products'
 
 ]
