+++ BackEnd/user/admin.py
@@ -29,19 +29,4 @@ class UserAdminConfig(UserAdmin):
     )
 
 
-class CategoryAdmin(admin.ModelAdmin):
-    prepopulated_fields = {'slug': ('title',)}
-    list_display = ("title",)
-    list_display_links = ("title",)
-
-
-class CityAdmin(admin.ModelAdmin):
-    prepopulated_fields = {'slug': ('title',)}
-    list_display = ("title",)
-    list_display_links = ("title",)
-
-
 admin.site.register(User, UserAdminConfig)
-admin.site.register(Category, CategoryAdmin)
-admin.site.register(City, CityAdmin)
-admin.site.register(Number)
