+++ BackEnd/user/models.py
@@ -54,42 +54,7 @@ class User(AbstractBaseUser, PermissionsMixin):
         return self.user_name
 
 
-class Category(models.Model):
-    title = models.CharField("Категорія", max_length=150)
-    description = models.TextField("Опис", blank=True)
-    slug = models.SlugField(max_length=160, unique=True)
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = "Категорія"
-        verbose_name_plural = "Категорії"
 
 
-class Number(models.Model):
-    number = models.CharField("Номер", max_length=50)
-
-    def __str__(self):
-        return self.number
-
-    class Meta:
-        verbose_name = "Номер"
-        verbose_name_plural = "Номери"
-
-
-class City(models.Model):
-    title = models.CharField("Місто", max_length=50)
-    slug = models.SlugField(max_length=160, unique=True)
-    number = models.ManyToManyField(
-        Number, verbose_name="Номер", null=True
-    )
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = "Місто"
-        verbose_name_plural = "Міста"
